refactor(connection): correct connection-manager file name
