+++ src/connections/connection-manager.ts
