+++ src/model/interfaces/create-model.interface.ts
@@ -1,5 +1,5 @@
 import { Schema } from '../../schema/schema';
-import { ConnectionManager } from '../../connections/connection-mannager';
+import { ConnectionManager } from '../../connections/connection-manager';
 
 export interface CreateModel {
   name: string;
